Create README.md
My name is Quiren Mungai and I am an administrative professional.  I’m seeking a certificate in cyber security and cyber security analyst. My interest in cyber security comes from a desire to know how to better protect myself and my family from the fast changing world of the internet.  As a cyber security analyst I can have the opportunity to analyze data and see what the current trends are relating to cyber security and how I can better guard myself and those around me. 

After studying the foundation of cybersecurity security I have a better understanding of  ways to secure data and also the programs that are used for the data security. 

Quiren Mungai